Modify ELO "k" factor
The elo ratings of most bots fluctuate by around 250 points even when
just playing constantly without being modified. In the IRC chat on
September 30 multiple users expressed support for lowering the k factor
to lower the volatility of the hero's elos. This changes the k factor
for experienced bots from 32 to 16 and rescales the k factor for new
bots so that it gradually descends from 100 to 16 rather than 100 to 32
like it used to before.

Based on some test code I wrote at http://pastebin.com/sX0uaqzK, if four
bots of exactly even strength played each other for one day (which is
around 420 games), then with a k factor of 32 their elo range
would be around 350 points, whereas with a k factor of 16 their elo
range is around 220 points. Since not all bots are of even strength on
Vindinium, this should be less in practice. Meanwhile, a k factor of 16
is still large enough to allow plently of upward mobility for a bot that
improves and starts winning more games than its rating predicts.